% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark \code{\link{rendr}} solvers}
\usage{
benchmark(
  N = 100,
  sys = rsys_examples("schlogl"),
  solver = ssa,
  progress = FALSE,
  parallel = FALSE,
  cores = detectCores(),
  ...
)
}
\arguments{
\item{N}{number of simulations (default 100)}

\item{sys}{an instance of either \code{\link{rsys}} or \code{\link{rdsys}}, must match \code{solver} type (well-stirred vs inhomogeneous) (default \code{rsys_examples('schlogl')})}

\item{solver}{one of 'ssa' (default), 'rre', 'issa', 'nsm'}

\item{progress}{if \code{TRUE} (default \code{FALSE}) display progress of completed runs}

\item{parallel}{if \code{TRUE} (default \code{FALSE}) runs the solver instances in parallel on \code{cores} number of CPU cores}

\item{cores}{number of cores to use if \code{parallel} is \code{TRUE}}

\item{...}{additional options to be passed to \code{solver}}
}
\description{
Benchmark \code{\link{rendr}} solvers
}
