% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compiles a reaction network to a c++ shared object.}
\usage{
compile(network, display_name = "")
}
\arguments{
\item{network}{network S3 object from chemnet::network function.}

\item{display_name}{name used to identify model. If left blank, a unique name will be automatically generated.}
}
\value{
theme object
}
\description{
Compiles a reaction network to a c++ shared object.
}
