% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issa.R
\name{issa}
\alias{issa}
\title{Inhomogeneous Stochastic Simulation Algorithm solver}
\usage{
issa(model)
}
\arguments{
\item{model}{an instance of the \code{rdmodel} class}
}
\value{
the solution to the system as a list
}
\description{
Inhomogeneous Stochastic Simulation Algorithm solver
}
