% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Parses a string containing a representation of a chemical network}
\usage{
network(string)
}
\arguments{
\item{string}{string containing the chemical network representation}
}
\value{
reaction network list object
}
\description{
Parses a string containing a representation of a chemical network
}
