% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsm.R
\name{nsm}
\alias{nsm}
\title{Next Subvolume Method implementation of an ISSA solver}
\usage{
nsm(sys, verbose = TRUE)
}
\arguments{
\item{sys}{an instance of the \code{rdsys} class}

\item{verbose}{controls if output is generated during during run (default TRUE)}
}
\value{
the solution to the system as a list
}
\description{
Next Subvolume Method implementation of an ISSA solver
}
