% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmodel.R
\name{rdmodel}
\alias{rdmodel}
\title{Reaction-diffusion network models}
\usage{
rdmodel(network, volume, D, T)
}
\arguments{
\item{network}{an instance of the \code{network} class from the \code{bondr} package}

\item{volume}{an instance of the \code{volume} class}

\item{D}{diffusion coefficients for each species in the order returned by \code{species(network)}}

\item{T}{simulation length}
}
\value{
an instance of the \code{rdmodel} class
}
\description{
Reaction-diffusion network models
}
