% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmodel.R
\name{rdmodel_examples}
\alias{rdmodel_examples}
\title{Predefined models for use with reaction-diffusion network solvers}
\usage{
rdmodel_examples(name)
}
\arguments{
\item{name}{model name, one of: schnakenberg, ...}
}
\value{
an instance of the \code{rdmodel} class
}
\description{
Predefined models for use with reaction-diffusion network solvers
}
