% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsol.R
\name{rdsol}
\alias{rdsol}
\title{Reaction-diffusion system solution}
\usage{
rdsol(sys, t, u)
}
\arguments{
\item{sys}{\code{\link{rdsys}} instance used to produce the solution}

\item{t}{\code{\link{vector}} of times in the solution}

\item{u}{\code{\link{list}} of \code{\link[tibble:tibble]{tibble::tibble}}s of solution states corresponding to each of the time points in \code{t}}
}
\value{
\code{\link{rdsol}} instance
}
\description{
Reaction-diffusion system solution
}
