% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsol.R
\name{rdsol_plot}
\alias{rdsol_plot}
\title{Reaction network solution plot}
\usage{
rdsol_plot(rsol)
}
\arguments{
\item{rsol}{a solution to a reaction-diffusion system}
}
\value{
a \code{ggplot2} plot of the solution
}
\description{
Reaction network solution plot
}
