% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsol.R
\name{rdsol_quantities}
\alias{rdsol_quantities}
\title{Aggregator for species quantities}
\usage{
rdsol_quantities(rdsol, average = FALSE, index = NULL)
}
\arguments{
\item{rdsol}{a solution to a reaction-diffusion system}

\item{index}{an optional voxel index to filter quantity extraction}
}
\value{
a tibble with a column for the solution time points, and a column for each species' quantities
}
\description{
Aggregator for species quantities
}
