% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsol.R
\name{rdsol_summarise}
\alias{rdsol_summarise}
\title{Aggregator for species quantities}
\usage{
rdsol_summarise(sol, func = sum, index = NULL)
}
\arguments{
\item{sol}{\code{\link{rdsol}} instance}

\item{func}{summary \code{\link{function}} to apply across voxels for each species at each time (default \code{\link{sum}})}

\item{index}{an optional voxel index to filter quantity extraction}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}} with a column for the solution time points, and a column for each species' quantities
}
\description{
Aggregator for species quantities
}
