% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsys.R
\name{rdsys}
\alias{rdsys}
\title{Reaction-diffusion network systems}
\usage{
rdsys(network, volume, D, T)
}
\arguments{
\item{network}{an instance of the \code{\link[bondr:parse_network]{bondr::network}} class}

\item{volume}{an instance of the \code{\link[spurcore:volume]{spurcore::volume}} class}

\item{D}{\code{\link{vector}} of diffusion coefficients for each species}

\item{T}{simulation length}
}
\value{
\code{\link{rdsys}} instance
}
\description{
Reaction-diffusion network systems
}
