% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsystem.R
\name{rdsystem_examples}
\alias{rdsystem_examples}
\title{Predefined systems for use with reaction-diffusion network solvers}
\usage{
rdsystem_examples(name)
}
\arguments{
\item{name}{system name, one of: schnakenberg, ...}
}
\value{
an instance of the \code{rdsys} class
}
\description{
Predefined systems for use with reaction-diffusion network solvers
}
