% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel.R
\name{rmodel_examples}
\alias{rmodel_examples}
\title{Predefined models for use with reaction network solvers}
\usage{
rmodel_examples(name)
}
\arguments{
\item{name}{model name, one of: mm, schlogl, gbk, pc}
}
\value{
an instance of the \code{rmodel} class
}
\description{
Predefined models for use with reaction network solvers
}
