% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsol.R
\name{rsol}
\alias{rsol}
\title{Reaction system solution}
\usage{
rsol(sys, sol)
}
\arguments{
\item{sys}{\code{\link{rsys}} instance used to produce the solution}

\item{sol}{\code{\link[tibble:tibble]{tibble::tibble}} solution produced by \code{\link{ssa}} or \code{\link{rre}}}
}
\value{
\code{\link{rsol}} instance
}
\description{
Reaction system solution
}
