% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsol.R
\name{rsol_plot}
\alias{rsol_plot}
\title{Reaction network solution plot}
\usage{
rsol_plot(rsol)
}
\arguments{
\item{rsol}{a solution to a well-stirred reaction system}
}
\value{
a \code{ggplot2} plot of the solution
}
\description{
Reaction network solution plot
}
