% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsolution.R
\name{rsolution_plot}
\alias{rsolution_plot}
\title{Reaction network solution plot}
\usage{
rsolution_plot(rsolution)
}
\arguments{
\item{solution}{a solution to a reaction system}
}
\value{
a \code{ggplot2} plot of the solution
}
\description{
Reaction network solution plot
}
