% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsys.R
\name{rsys_examples}
\alias{rsys_examples}
\title{Predefined systems for use with reaction network solvers}
\usage{
rsys_examples(name)
}
\arguments{
\item{name}{system name, one of: mm, schlogl, gbk, pc}
}
\value{
an instance of the \code{rsys} class
}
\description{
Predefined systems for use with reaction network solvers
}
