% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel.R
\name{rsystem}
\alias{rsystem}
\title{Reaction network systems}
\usage{
rsystem(network, state, T)
}
\arguments{
\item{network}{an instance of the \code{network} class from the \code{bondr} package}

\item{state}{vector containing the initial species quantities in the order returned by \code{species(network)}}

\item{T}{simulation length}
}
\value{
an instance of the \code{rsystem} class
}
\description{
Reaction network systems
}
