% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel.R
\name{rsystem_examples}
\alias{rsystem_examples}
\title{Predefined systems for use with reaction network solvers}
\usage{
rsystem_examples(name)
}
\arguments{
\item{name}{system name, one of: mm, schlogl, gbk, pc}
}
\value{
an instance of the \code{rsystem} class
}
\description{
Predefined systems for use with reaction network solvers
}
