% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{solution_plot}
\alias{solution_plot}
\title{Reaction network solution plot}
\usage{
solution_plot(solution)
}
\arguments{
\item{solution}{a solution to a reaction system}
}
\value{
a \code{ggplot2} plot of the solution
}
\description{
Reaction network solution plot
}
