% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{Generates a realization of the Chemical Master Equation solution using the Stochastic Simulation Algorithm}
\usage{
ssa(model, force_compile = FALSE)
}
\arguments{
\item{model}{an instance of the \code{rmodel} class}

\item{force_compile}{if set to \code{TRUE}, forces the overwriting and recompilation of the network source file}
}
\value{
the solution to the system as a \code{data.frame}
}
\description{
Generates a realization of the Chemical Master Equation solution using the Stochastic Simulation Algorithm
}
