% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa_trajest}
\alias{ssa_trajest}
\title{Optimised trajectory estimate (average trajectory) using SSA}
\usage{
ssa_trajest(
  sys,
  trajectories = 1,
  length.out = 100,
  k = NULL,
  force_compile = FALSE
)
}
\arguments{
\item{sys}{\code{\link{rsys}} instance}

\item{trajectories}{number of trajectories to generate}

\item{length.out}{length of solution output (table rows) (default 100)}

\item{k}{\code{\link{vector}} of reaction rates corresponding to the reactions in \code{sys}, overrides those contained if \code{sys} if provided}

\item{force_compile}{if set to \code{TRUE}, forces the overwriting and recompilation of the network source file}
}
\value{
named \code{\link{numeric}} vector or \code{\link{matrix}}, depending on \code{length.out}
}
\description{
Optimised trajectory estimate (average trajectory) using SSA
}
