% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauleap.R
\name{tauleap}
\alias{tauleap}
\title{Adaptive tau-leaping solver}
\usage{
tauleap(
  sys,
  length.out = 100,
  all.out = FALSE,
  trajectories = 1,
  parallel = FALSE,
  cores = detectCores(),
  average = FALSE,
  k = NULL,
  force_compile = FALSE,
  verbose = FALSE,
  method = "adaptive"
)
}
\arguments{
\item{sys}{\code{\link{rsys}} instance}

\item{length.out}{length of solution output (table rows) (default 100)}

\item{all.out}{if \code{TRUE} (default \code{FALSE}), ignore \code{length.out} and return entire solution}

\item{trajectories}{number of trajectories to generate}

\item{parallel}{if \code{TRUE} (default \code{FALSE}) generate trajectories using multiple CPU cores}

\item{cores}{number of cores to use if \code{parallel} is \code{TRUE} (default is all system cores)}

\item{average}{if \code{TRUE} (default \code{FALSE}) and generating multiple trajectories, averages trajectories at sample times; incompatible with \code{all.out = TRUE}}

\item{k}{\code{\link{numeric}} vector of reaction rates corresponding to the reactions in \code{sys}, overrides those contained if \code{sys} if provided}

\item{force_compile}{if set to \code{TRUE}, forces the overwriting and recompilation of the network source file}

\item{verbose}{print diagnostic information}

\item{method}{one of: 'adaptive' (default), 'implicit', 'explicit'}
}
\value{
\code{\link{rsol}} instance
}
\description{
Generates a realization of the solution to the Chemical Master Equation.
}
