% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume}
\alias{volume}
\title{Volume class for setting initial conditions of inhomogeneous systems}
\usage{
volume(dims, seed = numeric())
}
\description{
Volume class for setting initial conditions of inhomogeneous systems
}
