% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issa.R
\name{issa}
\alias{issa}
\title{Inhomogeneous Stochastic Simulation Algorithm solver}
\usage{
issa(sys, verbose = TRUE)
}
\arguments{
\item{sys}{an instance of the \code{rdsys} class}

\item{verbose}{controls if output is generated during during run (default TRUE)}
}
\value{
the solution to the system as a list
}
\description{
Inhomogeneous Stochastic Simulation Algorithm solver
}
