% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issa.R
\name{issa}
\alias{issa}
\title{Inhomogeneous Stochastic Simulation Algorithm (ISSA) solver}
\usage{
issa(sys, verbose = TRUE, force_compile = FALSE)
}
\arguments{
\item{sys}{an instance of the \code{\link{rdsys}} class}

\item{verbose}{controls if output is generated during during run (default \code{TRUE})}
}
\value{
Solution to the system as a \code{\link{list}}
}
\description{
Inhomogeneous Stochastic Simulation Algorithm (ISSA) solver
}
