% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issa.R
\name{nsm}
\alias{nsm}
\title{Next Subvolume Method (NSM) solver}
\usage{
nsm(sys, verbose = TRUE)
}
\arguments{
\item{sys}{an instance of the \code{\link{rdsys}} class}

\item{verbose}{controls if output is generated during during run (default \code{TRUE})}
}
\value{
Solution to the system as a \code{\link{list}}
}
\description{
Next Subvolume Method (NSM) solver
}
