% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsm.R
\name{nsm}
\alias{nsm}
\title{Next Subvolume Method implementation of an ISSA solver}
\usage{
nsm(model)
}
\arguments{
\item{model}{an instance of the \code{rdmodel} class}
}
\value{
the solution to the system as a list
}
\description{
Next Subvolume Method implementation of an ISSA solver
}
