% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issa.R
\name{nsm}
\alias{nsm}
\title{Next Subvolume Method (NSM) solver}
\usage{
nsm(
  sys,
  length.out = 100,
  all.out = FALSE,
  trajectories = 1,
  parallel = FALSE,
  cores = detectCores,
  average = FALSE,
  verbose = TRUE,
  k = NULL,
  force_compile = FALSE
)
}
\arguments{
\item{sys}{an instance of the \code{\link{rdsys}} class}

\item{length.out}{length of solution output (table rows) (default 100)}

\item{all.out}{if \code{TRUE} (default \code{FALSE}), ignore \code{length.out} and return entire solution}

\item{verbose}{controls if output is generated during during run (default \code{TRUE})}

\item{k}{\code{\link{numeric}} vector of reaction rates corresponding to the reactions in \code{sys}, overrides those contained if \code{sys} if provided}

\item{force_compile}{if \code{TRUE} (default \code{FALSE}), force a recompile of the reaction network}
}
\value{
\code{\link{rdsol}} instance
}
\description{
Next Subvolume Method (NSM) solver
}
