% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issa.R
\name{nsm}
\alias{nsm}
\title{Next Subvolume Method (NSM) solver}
\usage{
nsm(sys, verbose = TRUE, force_compile = FALSE)
}
\arguments{
\item{sys}{an instance of the \code{\link{rdsys}} class}

\item{verbose}{controls if output is generated during during run (default \code{TRUE})}

\item{force_compile}{if \code{TRUE} (default \code{FALSE}), force a recompile of the reaction network}
}
\value{
\code{\link{rdsol}} instance
}
\description{
Next Subvolume Method (NSM) solver
}
