% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsm.R
\name{nsm}
\alias{nsm}
\title{Next Subvolume Method implementation of an ISSA solver}
\usage{
nsm(network, D, volume, tspan)
}
\arguments{
\item{network}{a reaction network object created using \code{parse_network} from the \code{bondr} package}

\item{D}{vector of diffusion constants for each species, in the order matching \code{species(network)}}

\item{volume}{an instance of the \code{volume} S3 class}

\item{tspan}{the simulation time}
}
\value{
the solution to the system as a list
}
\description{
Next Subvolume Method implementation of an ISSA solver
}
