% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdmodel.R
\name{rdmodel}
\alias{rdmodel}
\title{Reaction-diffusion network models}
\usage{
rdmodel(network, volume, D, tspan)
}
\arguments{
\item{network}{an instance of the \code{network} class from the \code{bondr} package}

\item{volume}{an instance of the \code{volume} class}

\item{D}{diffusion coefficients for each species in the order returned by \code{species(network)}}

\item{tspan}{vector containing the start and stop times for simulation}
}
\value{

}
\description{
Reaction-diffusion network models
}
