% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsol.R
\name{rdsol_quantities}
\alias{rdsol_quantities}
\title{Aggregator for species quantities}
\usage{
rdsol_quantities(sol, average = FALSE, index = NULL)
}
\arguments{
\item{sol}{\code{\link{rdsol}} instance}

\item{average}{if \code{TRUE}, will return quantities averages over the number of voxels at each time}

\item{index}{an optional voxel index to filter quantity extraction}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble}} with a column for the solution time points, and a column for each species' quantities
}
\description{
Aggregator for species quantities
}
