% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsys.R
\name{rdsys_examples}
\alias{rdsys_examples}
\title{Example reaction-diffusion systems}
\usage{
rdsys_examples(name = NULL)
}
\arguments{
\item{name}{one of: 'schnakenberg'}
}
\value{
an instance of the \code{\link{rdsys}} class
}
\description{
For use in \code{\link{issa}} / \code{\link{nsm}}.
}
