% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmodel.R
\name{rmodel}
\alias{rmodel}
\title{Reaction network models}
\usage{
rmodel(network, state, tspan)
}
\arguments{
\item{network}{an instance of the \code{network} class from the \code{bondr} package}

\item{state}{vector containing the initial species quantities in the order returned by \code{species(network)}}

\item{tspan}{vector containing the start and stop times for simulation}
}
\value{

}
\description{
Reaction network models
}
