% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rre.R
\name{rre}
\alias{rre}
\title{Solves a reaction system deterministically using the Reaction Rate Equation}
\usage{
rre(network, y, tspan)
}
\arguments{
\item{network}{a reaction network object created using \code{parse_network} from the \code{chemnet} package}

\item{y}{the starting state of the system at the initial time}

\item{tspan}{the simulation time}
}
\value{
the solution to the system as a \code{tibble}
}
\description{
Solves a reaction system deterministically using the Reaction Rate Equation
}
