% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rre.R
\name{rre}
\alias{rre}
\title{Reaction rate equation solver}
\usage{
rre(sys)
}
\arguments{
\item{sys}{\code{\link{rsys}} instance}
}
\value{
\code{\link{rsol}} instance
}
\description{
Deterministic solver for reaction systems.
}
