% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rre.R
\name{rre}
\alias{rre}
\title{Reaction rate equation solver}
\usage{
rre(sys, length.out = 100)
}
\arguments{
\item{sys}{\code{\link{rsys}} instance}

\item{length.out}{length of solution output (table rows) (default 100)}
}
\value{
\code{\link{rsol}} instance
}
\description{
Deterministic solver for reaction systems.
}
