% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rre.R
\name{rre}
\alias{rre}
\title{Solves a reaction system deterministically using the Reaction Rate Equation}
\usage{
rre(sys)
}
\arguments{
\item{sys}{an instance of the \code{rsys} class}
}
\value{
the solution to the system as a \code{tibble}
}
\description{
Solves a reaction system deterministically using the Reaction Rate Equation
}
