% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsol.R
\name{rsol_plot}
\alias{rsol_plot}
\title{Reaction network solution plot}
\usage{
rsol_plot(rsol)
}
\arguments{
\item{rsol}{solution to a well-stirred reaction system}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot}} plot of the solution
}
\description{
Reaction network solution plot
}
