% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsys.R
\name{rsys}
\alias{rsys}
\title{Reaction systems}
\usage{
rsys(network, state, T)
}
\arguments{
\item{network}{an instance of the \code{\link[bondr:network]{bondr::network}}}

\item{state}{\code{\link{vector}} containing the initial species quantities in the order returned by \code{species(network)}}

\item{T}{simulation length}
}
\value{
\code{\link{rsys}} instance
}
\description{
Reaction systems
}
