% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsys.R
\name{rsys_examples}
\alias{rsys_examples}
\title{Example reaction systems}
\usage{
rsys_examples(name = NULL)
}
\arguments{
\item{name}{one of: \code{'mm'} (default, Michaelis Menten), or \code{'shlogl'}}
}
\value{
\code{\link{rsys}} instance
}
\description{
For use with \code{\link{rre}} / \code{\link{ssa}}.
}
