% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{Stochastic simulation algorithm solver}
\usage{
ssa(sys, k = NULL, record_all = TRUE, force_compile = FALSE)
}
\arguments{
\item{sys}{\code{\link{rsys}} instance}

\item{k}{\code{\link{vector}} of reaction rates corresponding to the reactions in \code{sys}, overrides those contained if \code{sys} if provided}

\item{record_all}{if \code{TRUE} (default), record the system state at all time steps}

\item{force_compile}{if set to \code{TRUE}, forces the overwriting and recompilation of the network source file}
}
\value{
\code{\link{rsol}} instance
}
\description{
Generates a realization of the solution to the Chemical Master Equation.
}
